refactor: optimized images, converted from svg to webp
