+++ src/components/missionPage/missionPage.module.css
@@ -12,6 +12,3 @@
   background-color: var(--white);
   /* color: black; */
 }
-.body > p {
-  color: black;
-}
